requirements.txt
absl-py==2.2.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
ccimport==0.4.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.2
cumm-cu118==0.7.11
cycler==0.12.1
dash==3.0.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
einops==0.8.1
entrypoints==0.4
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gradio==4.29.0
gradio_client==0.16.1
gradio_imageslider==0.0.20
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipycanvas==0.13.3
ipyevents==2.0.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kaolin==0.17.0
kiwisolver==1.4.8
lark==1.2.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multimethod==2.0
mypy_extensions==1.1.0
narwhals==1.36.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pccm==0.4.16
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
plyfile==1.1
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pybind11==2.13.6
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygltflib==1.16.4
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
segmentation_models_pytorch==0.5.0
semantic-version==2.10.0
SharedArray==3.2.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spconv-cu118==2.3.8
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.0
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
threadpoolctl==3.6.0
timm==1.0.15
toml==0.10.2
tomlkit==0.12.0
torch==2.2.0+cu118
torch-geometric==2.6.1
torch_cluster==1.6.3+pt22cu118
torch_scatter==2.1.2+pt22cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.2.0
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
usd-core==25.5.1
uvicorn==0.34.2
warp-lang==1.7.1
watchdog==6.0.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
yapf==0.43.0
yarl==1.20.0
zipp==3.21.0
